+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## future
+- Support sixels
+
 ## v2.0.0
 - Rename `TermText::to_string` to `to_string_cache`.